+++ assets/src/ba_data/python/ba/_plugin.py
@@ -117,9 +117,10 @@ class PluginSubsystem:
                      subs=[('${NUM}', str(len(disappeared_plugs)))]),
                 color=(1, 1, 0),
             )
+            plugnames = ', '.join(disappeared_plugs)
             _ba.log(
                 f'{len(disappeared_plugs)} plugin(s) no longer found:'
-                f' {str(disappeared_plugs)[1:-1]}.',
+                f' {plugnames}.',
                 to_server=False)
             for goneplug in disappeared_plugs:
                 del _ba.app.config['Plugins'][goneplug]