+++ src/ballistica/ballistica.cc
@@ -21,7 +21,7 @@
 namespace ballistica {
 
 // These are set automatically via script; don't modify them here.
-const int kAppBuildNumber = 20638;
+const int kAppBuildNumber = 20641;
 const char* kAppVersion = "1.7.4";
 
 // Our standalone globals.